added the http server upload pack
